refactor: 패키지 이동 및, REST API URI 수정
