+++ src/main/java/chess/entity/ChessGameEntity.java
@@ -1,4 +1,4 @@
-package chess.repository.entity;
+package chess.entity;
 
 import chess.domain.game.ChessGame;
