+++ src/main/java/chess/service/SparkChessService.java
@@ -4,10 +4,10 @@ import chess.domain.game.ChessGame;
 import chess.domain.game.GameSwitch;
 import chess.domain.game.Turn;
 import chess.domain.piece.Team;
-import chess.repository.entity.BoardEntity;
-import chess.repository.entity.ChessGameEntity;
-import chess.repository.spark.BoardDao;
-import chess.repository.spark.ChessGameDao;
+import chess.entity.BoardEntity;
+import chess.entity.ChessGameEntity;
+import chess.dao.spark.BoardDao;
+import chess.dao.spark.ChessGameDao;
 import chess.service.util.BoardEntitiesToBoardConvertor;
 import java.util.ArrayList;
 import java.util.List;
